put existing apis into 'v1' namespace
